refactor: move NFS constants to a dedicated module
Resolves: #139
