+++ src/client/arguments/nfsv3/create.rs
@@ -1,6 +1,6 @@
 use std::io::{Result, Write};
 
-use crate::nfsv3::NFS3_CREATEVERFSIZE;
+use crate::consts::nfsv3::NFS3_CREATEVERFSIZE;
 use crate::serializer::files::dir_op_arg;
 use crate::serializer::{array, u32};
 use crate::vfs::create::{Args, How, HowMode, Verifier};
