+++ src/lib.rs
@@ -2,9 +2,9 @@
 
 pub mod allocator;
 pub mod client;
+pub mod consts;
 mod context;
 pub mod mount;
-pub mod nfsv3;
 pub mod parser;
 pub mod rpc;
 pub mod serializer;
