+++ src/serializer/mount/export.rs
@@ -1,8 +1,8 @@
 use std::io;
 use std::io::Write;
 
-use crate::mount;
-use crate::mount::{export, MOUNT_HOST_NAME_LEN};
+use crate::consts::mount::MOUNT_HOST_NAME_LEN;
+use crate::mount::{self, export};
 use crate::serializer::files::file_path;
 use crate::serializer::{bool, string_max_size};
